thesis/requirements.txt
flask
python-dotenv
boto3
Pillow
google-cloud-vision
gtts
pinecone
sentence_transformers
face_recognition
opencv-python
moviepy
requests
pronouncing
openai
mediapipe
torch
face-alignment
librosa
google-cloud-texttospeech
imageio-ffmpeg
Flask-SQLAlchemy
flask_login
Flask-Migrate
groq